+++ herbs/image_view.py
@@ -215,7 +215,7 @@ class ImageView(QObject):
     def set_data(self, image_file):
         if image_file.n_channels > self.max_num_channels:
             raise ValueError(
-                'HERBS supports at most {} image channels.'.format(
+                'DriftlessMap supports at most {} image channels.'.format(
                     self.max_num_channels
                 )
             )
@@ -644,4 +644,3 @@ class ImageView(QObject):
         self.get_corner_and_lines()
 
     # def clear_curve_widget(self):
-